requirements.txt
astroid==2.0.1
autopep8==1.3.5
certifi==2018.4.16
chardet==3.0.4
coverage==4.5.1
coveralls==1.3.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.6
django-cors-middleware==1.3.1
django-extensions==2.1.0
django-heroku==0.3.1
django-js-asset==1.1.0
django-mptt==0.9.1
djangorestframework==3.8.2
docopt==0.6.2
gunicorn==19.9.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==2.1.0
psycopg2==2.7.5
pycodestyle==2.4.0
PyJWT==1.6.4
pylint==2.0.1
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.10.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
typed-ast==1.1.0
urllib3==1.23
whitenoise==3.3.1
wrapt==1.10.11
astroid==2.0.1
autopep8==1.3.5
certifi==2018.4.16
chardet==3.0.4
coverage==4.5.1
coveralls==1.3.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.6
django-cors-middleware==1.3.1
django-extensions==2.1.0
django-heroku==0.3.1
django-js-asset==1.1.0
django-model-utils==3.1.2
django-mptt==0.9.1
django-notifications-hq==1.5.0
djangorestframework==3.8.2
docopt==0.6.2
gunicorn==19.9.0
idna==2.7
isort==4.3.4
jsonfield==2.0.2
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==2.1.0
psycopg2==2.7.5
pycodestyle==2.4.0
PyJWT==1.6.4
pylint==2.0.1
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.10.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
typed-ast==1.1.0
urllib3==1.23
whitenoise==3.3.1
wrapt==1.10.11
